+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [0.9.4] - 2024-05-20
+
+### Added
+
+- Output operational presence errors at end
+- Logging functions for consistent output format
+
 ## [0.9.4] - 2024-05-17
 
 ### Added